requirements.txt
torch
transformers
sentence-transformers
Pillow
gradio
torchvision
